+++ MOP/src/GameObjects/Items/CashRegisterHook.cs
@@ -78,18 +78,16 @@ namespace MOP
                     items[i].AddComponent<ItemHook>();
 
                     // Hook the TriggerMinorObjectRefresh to Confirm and Spawn all actions
-                    if (items[i].name.Contains("shopping bag"))
+                    if (items[i].name.Equals("shopping bag(itemx)"))
                     {
                         FsmHook.FsmInject(items[i], "Confirm", TriggerMinorObjectRefresh);
                         FsmHook.FsmInject(items[i], "Spawn all", TriggerMinorObjectRefresh);
                     }
-
-                    if (items[i].name.StartsWith("spark plug box"))
+                    else if (items[i].name.EqualsAny("spark plug box(Clone)", "car light bulb box(Clone)"))
                     {
                         FsmHook.FsmInject(items[i], "Create Plug", WipeUseLoadOnSparkPlugs);
                     }
-
-                    if (items[i].name.ContainsAny("alternator belt(Clone)", "oil filter(Clone)", "battery(Clone)"))
+                    else if (items[i].name.EqualsAny("alternator belt(Clone)", "oil filter(Clone)", "battery(Clone)"))
                     {
                         PlayMakerFSM fanbeltUse = items[i].GetPlayMakerByName("Use");
                         FsmState loadFanbelt = fanbeltUse.FindFsmState("Load");