+++ MOP/src/GameObjects/Others/SatsumaInAreaCheck.cs
@@ -24,7 +24,7 @@ namespace MOP
         // Attaches to gameobject. This script checks if the Satsuma is in the object's area,
         // and sends the info to it about it.
 
-        const string ReferenceItem = "gearbox";
+        const string ReferenceItem = "car body";
         
         // Trigger collider
         BoxCollider collider;
@@ -44,7 +44,6 @@ namespace MOP
                 Satsuma.instance.IsSatsumaInInspectionArea = true;
                 if (this.gameObject.name == "MOP_ParcFermeTrigger")
                 {
-                    MSCLoader.ModConsole.Print("Satsuma in Parc Ferme");
                     Satsuma.instance.IsSatsumaInParcFerme = true;
                 }
             }
@@ -57,7 +56,6 @@ namespace MOP
                 Satsuma.instance.IsSatsumaInInspectionArea = false;
                 if (this.gameObject.name == "MOP_ParcFermeTrigger")
                 {
-                    MSCLoader.ModConsole.Print("Satsuma left Parc Ferme");
                     Satsuma.instance.IsSatsumaInParcFerme = false;
                 }
             }